Fix build errors by deleting info.json
I got a build error about the keys in the keymap not agreeing with the
local `info.json` (which was copied over from the ergodox-ez board).
I thought that if the error is that they disagree, then if I delete the
`info.json`, they can't disagree. Turns out that worked! Hopefully the
only purpose of `info.json` is documentation.